+++ wp-content/plugins/advisories-headers/spec/headers.spec.php
@@ -10,7 +10,7 @@ describe(\Dxw\AdvisoriesHeaders\Headers::class, function () {
 			it('it registers filters for wp_headers without a CSP', function () {
 				allow('is_admin')->toBeCalled()->andReturn(true);
 				allow('add_filter')->toBeCalled();
-				expect('add_filter')->toBeCalled()->times(4);
+				expect('add_filter')->toBeCalled()->times(5);
 				expect('add_filter')->toBeCalled()->once()->with(
 					'wp_headers',
 					[$this->headers, 'addCacheControl']
@@ -29,6 +29,11 @@ describe(\Dxw\AdvisoriesHeaders\Headers::class, function () {
 					[$this->headers, 'addCSPScriptAttributes'],
 					99999
 				);
+				expect('add_filter')->toBeCalled()->once()->with(
+					'style_loader_tag',
+					[$this->headers, 'addCSPStyleAttributes'],
+					99999
+				);
 				$this->headers->register();
 				expect(true)->toBeTruthy();
 			});
@@ -37,7 +42,7 @@ describe(\Dxw\AdvisoriesHeaders\Headers::class, function () {
 			it('it registers filters for wp_headers including a CSP', function () {
 				allow('is_admin')->toBeCalled()->andReturn(false);
 				allow('add_filter')->toBeCalled();
-				expect('add_filter')->toBeCalled()->times(5);
+				expect('add_filter')->toBeCalled()->times(6);
 				expect('add_filter')->toBeCalled()->once()->with(
 					'wp_headers',
 					[$this->headers, 'addCacheControl']
@@ -83,6 +88,35 @@ describe(\Dxw\AdvisoriesHeaders\Headers::class, function () {
 				$input = ['src' => 'https://example.com', 'foo' => 'bar'];
 				$expected = ['src' => 'https://example.com', 'nonce' => '012345', 'foo' => 'bar'];
 				$result = $this->headers->addCSPScriptAttributes($input);
+				expect($result)->toEqual($expected);
+			});
+		});
+	});
+
+	describe('addCSPStyleAttributes->()', function () {
+		context('the tag is not a style tag', function () {
+			it('does nothing', function () {
+				$input = '<link rel="stylesheet" href="https://example.com"';
+				$result = $this->headers->addCSPStyleAttributes($input);
+				expect($result)->toEqual($input);
+			});
+		});
+		context('the tag is a style tag and a nonce is already set', function () {
+			it('does nothing', function () {
+				$input = '<style nonce="012345" foo="bar">';
+				$result = $this->headers->addCSPStyleAttributes($input);
+				expect($result)->toEqual($input);
+			});
+		});
+		context('the tag is a style tag and a nonce is not set', function () {
+			it('adds a nonce', function () {
+				allow('wp_create_nonce')->toBeCalled()->with('csp')->andReturn('012345');
+				allow('esc_attr')->toBeCalled()->andRun(function ($val) {
+					return $val;
+				});
+				$input = '<style foo="bar">';
+				$expected = '<style nonce="012345" foo="bar">';
+				$result = $this->headers->addCSPStyleAttributes($input);
 				expect($result)->toEqual($expected);
 			});
 		});
@@ -196,7 +230,7 @@ describe(\Dxw\AdvisoriesHeaders\Headers::class, function () {
 				$policy = "default-src 'self'; script-src 'self' 'nonce-012345' data: https://plausible.io ";
 				$policy .= "https://wordpress.org; connect-src 'self' data: https://plausible.io https://wordpress.org; ";
 				$policy .= "img-src 'self' data: https://plausible.io https://wordpress.org https://secure.gravatar.com; style-src 'self' ";
-				$policy .= "'unsafe-inline'; font-src 'self' data: https://wordpress.org; object-src 'none'; media-src ";
+				$policy .= "'nonce-012345'; font-src 'self' data: https://wordpress.org; object-src 'none'; media-src ";
 				$policy .= "'none'; frame-src 'none'; child-src 'none'; worker-src 'none'; manifest-src 'self'; ";
 				$policy .= "base-uri 'self'; form-action 'self'; frame-ancestors 'none';";
 				$expected = ['Content-Security-Policy' => $policy];